Created ClassicalBoard
